requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
hdbscan>=0.8.33

# Explainable AI
shap>=0.42.0
google-generativeai>=0.3.0

# Visualization (minimal - only for Streamlit charts)
plotly>=5.17.0

# Log parsing
python-evtx>=0.8.1

# Utilities
joblib>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0

